refactor(number): 🏗  Reorganize schema  modules groups
