+++ packages/number/source/schema/groups/finite/finite.test.ts
