+++ packages/number/source/schema/groups/infinity/infinity.test.ts
