+++ packages/number/source/schema/groups/infinity/infinity.ts
