+++ packages/number/source/schema/groups/nan/nan.test.ts
