+++ packages/number/source/schema/groups/nan/nan.ts
