+++ packages/number/source/schema/groups/negative/negative.ts
