+++ packages/number/source/schema/groups/safe/safe.ts
