+++ packages/number/source/schema/schema.ts
@@ -1,11 +1,11 @@
-export * from "./groups/finite.ts";
-export * from "./groups/float.ts";
-export * from "./groups/infinity.ts";
-export * from "./groups/integer.ts";
-export * from "./groups/multiple-of.ts";
-export * from "./groups/nan.ts";
-export * from "./groups/negative.ts";
-export * from "./groups/number.ts";
-export * from "./groups/positive.ts";
-export * from "./groups/safe.ts";
-export * from "./groups/zero.ts";
+export * from "./groups/finite/finite.ts";
+export * from "./groups/float/float.ts";
+export * from "./groups/infinity/infinity.ts";
+export * from "./groups/integer/integer.ts";
+export * from "./groups/multiple-of/multiple-of.ts";
+export * from "./groups/nan/nan.ts";
+export * from "./groups/negative/negative.ts";
+export * from "./groups/number/number.ts";
+export * from "./groups/positive/positive.ts";
+export * from "./groups/safe/safe.ts";
+export * from "./groups/zero/zero.ts";
